Automatically retrigger rvpkg on package changes (#161)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing of the related
issue when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.

Adds a dependency to the RV Packages to the step that installs the
package to improve the developer experience.

If someone is working on a package and modifying the files, re-building
will build a new rvpkg and install it in the compiled RV.

### Describe the reason for the change.

The developer experience was terrible. We had to install the package
manually. Now, when we change a python file and build, it does what
needs to be done to have it installed for the next debugging session.

### Describe what you have tested and on which operating system.

It's working on MacOS Ventura, but it's pretty platform-independent.

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>