Added RHEL vars for Redhat distros in CMake (#164)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added the possibility to generate a variable in CMake (on Linux only)
that stores the RHEL major release, to use in if statements in case of
problems between distro versions.

### Describe the reason for the change.
When running rvbootstrap and rvbuild, some libraries would be
'non-existant', even though they did exist; the problem was in the
library suffix which was different with what was needed (eg: .1.so
instead of .so.1). This problem would cause the commands to constantly
try to rebuild the 'missing' libraries, while there were no changes
brought to the files.

### Describe what you have tested and on which operating system.
Rocky 9.1
The problem occurred with libgc (in gc.cmake) and libcrypto (in
openssl.cmake)
### Add a list of changes, and note any that might need special
attention during the review.
- gc.cmake
- openssl.cmake
Both cmake files now have an if statement to check their RHEL value, and
change the suffix order accordingly.
- rv_targets.cmake
### If possible, provide screenshots.

---------

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>