Fix Linux install (#171)
[Fix Linux install]

A few of the PySide libraries generated on CentOS7 are not able to be
stripped because they are too large for strip to handle. To workaround
this issue, I simply changed the strip error from fatal to a warning.

I also added an additional mime type for fonts that strip was failing on
to the ignorelist.

Tested on CentOS7

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>