Issue-169: [Bug]: OpenRV fails to rebuild on Windows (#174)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

Fixes #169

Fixing build error that happens during a rebuild for Python/PySide

Tested rebuilding on Windows 11.

attention during the review.

Fixing FileExistsError by removing the 'CMakeLists.txt.old' before
renaming the current CMakeList file

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>