Add the cache keyword to the target vars (#186)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added CACHE INTERNAL "" to the platform targets (DARWIN, LINUX, WINDOWS)
### Describe the reason for the change.
These variables weren't exposed in RV Commercial, and they will be
needed (in the install hooks, for example)
### Describe what you have tested and on which operating system.

### Add a list of changes, and note any that might need special
attention during the review.
- rv_targets.cmake
### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>