Properly install RV_PACKAGE_LIST with rvpkg at the end of the build (#195)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Summarize your change.

In #161, I introduced a change to clean the RV_PACKAGE_LIST at each
configure stage so that if we update a package version, we don't have
the old package in the RV_PACKAGE_LIST. This behaviour is buggy, so I
prefer to revert that bit and try again another time with another
approach.

### Describe the reason for the change.

It breaks the rvinstall target that installs all the packages. 

### Describe what you have tested and on which operating system.

I tested on MacOS Ventura, but it's pretty much platform agnostic. 

### Add a list of changes and note any needing special attention during
the review.

N/A

### If possible, provide screenshots.

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>