Properly install RV_PACKAGE_LIST with rvpkg at the end of the build (#195)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

In #161, I introduced a change to clean the RV_PACKAGE_LIST at each
configure stage so that if we update a package version, we don't have
the old package in the RV_PACKAGE_LIST. This behaviour is buggy, so I
prefer to revert that bit and try again another time with another
approach.

It breaks the rvinstall target that installs all the packages.

I tested on MacOS Ventura, but it's pretty much platform agnostic.

the review.

N/A

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>