Disable OpenSSL asm runtimes on Windows (#197)
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>

<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

Reverts #150 on Windows.

There are reports in the field that there are build failures related to
nasm on the OpenSSL build. It seems to work fine in my test environment,
but not everywhere. It's safe to disable it on Windows since it was
disabled at first. It only got enabled because it was required on MacOS
and I wanted to have the same configuration on all 3 platforms.

It's a revert. It's the configuration that is being used in the
Commercial RV build.

attention during the review.

N/A

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>