Rewrite the rvinstall custom command (#198)
Moves the rvpkg installation commmand inside of a python script so the
package list doesn't have to be known when the custom command get
defined.

<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

Rewrite the step that installs the rvpkg packages so it always install
all the packages that are built into the stage/packages folder.

Before this change, the package targets that get defined after the
rvinstall target are not automatically installed into OpenRV. It's the
case when the OpenRV build is used as a sub-project, like when Autodesk
RV is built.

The script is platform agnostic, and tested on MacOS.

attention during the review.

N/A

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>