Package installation fix (#201)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

Removed unused source_icons folder in the annotate pacakge
Changed package installation code to maintain directory structure for
auxiliary folders
Change package loading to code to allow the required section in a
plugin's PACKAGE file to only need the package name and not version
number
Added some error checking to auxfile support code

MacOS, Windows

attention during the review.

---------

Signed-off-by: shannonyearwood <syearwood1059@gmail.com>
Signed-off-by: shannonyearwood <87450032+shannonyearwood@users.noreply.github.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>