Update OpenSSL, Python and PySide2 (#199)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

Updates the version to build of OpenSSL, Python and PySide2 (Qt for
Python).

Periodic maintenance task.

Building fine on my MacBook and on Autodesk's internal CI system.

attention during the review.

N/A

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>