Add rvio software support on Linux (#168)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added some new packages that install MesaGL libraries on the Linux build
for the rvio_sw binary.
### Describe the reason for the change.
In case any linux machine does not have a GPU, in which case the rvio
binary would not work. This is where rvio_sw comes in to save the day.
### Describe what you have tested and on which operating system.
Apparently would only be needed on Redhat distros, so the new binary and
libraries only get installed on those distros (Rocky, Centos, etc.)
### Add a list of changes, and note any that might need special
attention during the review.
New rvio_sw folder in imgtools
New CMake build libraries used in the rvio_sw binary:
- TwkGLFMesa
- MovieRV
Brought back the use of the RVIO_HW macro in rvio/main.cpp (from
shotgun-rv)
### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>