[BUG] Fix the color inspector final color mode (#179)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

The color inspector will now receive a 'graph-state-change' event
whenever the pixel color mode changes (the event already existed, it was
just about binding it). We use that event to trigger the correct code
inside the inspector.mu to update the inspector.

I also found another bug while working on this one, about the average
color pixel; instead of taking the pixel color from the current mode, it
would take the source pixel color no matter what.

Also fixed some janky inspector resetting event.
When in Final Color mode in the Color Inspector, a change in the graph
would not always update the inspector. You'd need to move the inspector
manually to update it.
Mac
attention during the review.
- inspector.mu

1. Download
[color_inspector_bug.rv.zip](https://github.com/AcademySoftwareFoundation/OpenRV/files/11772217/color_inspector_bug.rv.zip)
2. Load it up in RV of your choice.
3. Press F5 to bring up color inspector and in the little drop-down
triangle change menu to Final Rendered Color.
4. Go to View -> Rec709.
5. Note that inspector doesn't change.
6. Jiggle inspector, it should update.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>