Update OCIO v2 feature branch (#202)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.

Updating the feature branch before merging to get a cleaner diff on
GitHub

### Describe the reason for the change.

Get a cleaner diff

### Describe what you have tested and on which operating system.

n/a

### Add a list of changes, and note any that might need special
attention during the review.

I used a simple merge command

### If possible, provide screenshots.

---------

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Brian R Hanke <59420805+BrianHanke@users.noreply.github.com>
Signed-off-by: ThomBern <thomas.bernier@autodesk.com>
Signed-off-by: shannonyearwood <syearwood1059@gmail.com>
Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: shannonyearwood <87450032+shannonyearwood@users.noreply.github.com>
Co-authored-by: Bernard Laberge <117092886+bernie-laberge@users.noreply.github.com>
Co-authored-by: Brian R Hanke <59420805+BrianHanke@users.noreply.github.com>
Co-authored-by: shannonyearwood <87450032+shannonyearwood@users.noreply.github.com>
Co-authored-by: Thomas Bernier <42812610+ThomBern@users.noreply.github.com>
Co-authored-by: Kerby Geffrard <kerby.geffrard@autodesk.com>