Update documentation for Windows and Linux and build fix for RHEL8 (#208)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.

Updating Windows build instructions & Fixing Linux build issue

### Describe the reason for the change.

Using MSYS to build on Windows has many non obvious choices such as how
the PATH is set which is used to select App version; if the wrong
version of an App is used, the build will break.

### Describe what you have tested and on which operating system.

Win 10 & Win 11 and Linux Rocky 8

### Add a list of changes, and note any that might need special
attention during the review.

Most of the updates are for build docs but for Rocky8 there's a tweak on
the build code; I can remove it if it isn't wanted or incorrect.t
Re-adding the Py7zr workaround for PIP
Adding new file for Rocky8; it's the same as Rocky 9 except the OpenSSL
section which is new and only for Rocky8.


### If possible, provide screenshots.

---------

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>