Added new README header (#207)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Adding new badges and the OpenRV logo to the README file similar to
[OpenTimelineIO](https://github.com/AcademySoftwareFoundation/OpenTimelineIO/blob/main/README.md)

### Describe the reason for the change.
It looks cool

### Describe what you have tested and on which operating system.
Should not need testing on any precise platform, since it's only the
README
### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.
Basing it off of OpenTimelineIO
<img width="1052" alt="Screenshot 2023-07-20 at 3 31 41 PM"
src="https://github.com/AcademySoftwareFoundation/OpenRV/assets/42812610/e33b1e99-6a2d-47eb-9cf5-d1ecbcc3b57a">

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>