Added hash and branch to cache (#211)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added the 'CACHE STRING' flag to the RV_GIT_COMMIT_SHORT_HASH and
RV_GIT_BRANCH cmake variables to expose them in RV Commercial.
### Describe the reason for the change.
We want the correct commit info for the about box (although would only
be shown on a dev build)

### Describe what you have tested and on which operating system.
MacOS, should be good on other OS though.

### Add a list of changes, and note any that might need special
attention during the review.
- Making sure the CMakeCache is OK.
N.B.: might need to delete and recreate the CMakeCache.
### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>