Added RV Icon as a cmake var (#212)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added a variable in CMake that contains the path to the Open RV Icon.
Will be used to set the RV Commercial path.

### Describe the reason for the change.
Needed for the RV Commercial About Box and the Application icon.

### Describe what you have tested and on which operating system.
MacOS

### Add a list of changes, and note any that might need special
attention during the review.
 - RvCommon's CMakeLists
 - rv_global.cmake (to expose the ressource directory for RV Commercial)
 - RvApplication.cpp and RvDocument.cpp now use the CMake Icon var.

### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>