MIO-FFMPEG: Update mio-ffmpeg
Moving all the test in a subfolder in order to not mix the mio_ffmeg
build files with the external build files to avoid confusion when looking
at the build result and build intermediate files.
Adding a dependency to make it obvious that mio_ffmpeg doesn't build if
FFMPEG is not built first.

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>