Add a file logger system (with rotation and asynchronicity) (#176)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345

Fixes #54345
-->

This PR adds a new library, spdlog, which we will be using for file
logging.
A new library means a new dependence, and so there's a spdlog.cmake file
that automatically fetches and builds the library, cool stuff.
The library is used in a new file, FileLogger.cpp/.h, where the magic
happens.

This PR is a solution to a ticket, so users can share logs easily with
support.

Mac, Windows.
Builds on Rocky 9 (but not tested)

attention during the review.
- spldlog.cmake
- The file logging happens in FileLogger.cpp, called in
RvConsoleWindow.cpp.
<img width="769" alt="Screenshot 2023-06-13 at 1 52 04 PM"
src="https://github.com/AcademySoftwareFoundation/OpenRV/assets/42812610/178a03c5-fec6-48c5-9bd5-05edd17a2526">

---------

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>