Fix nested sequence OTIO export (#223)
[ 214: Fix OTIO Export with nested sequences ]

Fixes #214 

This branch fixes exporting nested sequences to OTIO. The problem was
with the kwargs parameters to `create_otio_from_rv_node` which
dispatches the call to a function for each supported node type. The
method to parse a track was setting parameters that only applied to the
`create_item` function so if there was any other node type (in this case
a second sequence) the call would fail.

To fix it, I simply just use the values from args/kwargs directly,
instead of mixing args with positional aruments and kwargs with named
arguments. This way both functions in the map have the same signature so
they can be called with the same parameters.

I noticed the same problematic pattern in a few other methods as well,
so I clean it up everywhere I found it. In these other cases its likely
impossible to run into the problem due to the topology of the graph (for
example we shouldn't ever have nested transitions), but the code is now
generic and independent of inputs node types so we can support
unexpected/unanticipated topologies.

Tested on OSX with the rep steps provided in the linked issue. Also
ensured that the arguments were still correctly being passed to the
hooks.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>