Revert "Re-enable mio-ffmpeg external build, now with CMake (#206)" (#227)
This reverts commit 3b509ab0c07c803aac4b9fbd406043d0adc21535.

<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

Fixes #226 

### Summarize your change.

Reverts the external mio-ffmpeg pull request

### Describe the reason for the change.

There's reports of build issues caused by the pull request being
reverted

### Describe what you have tested and on which operating system.

All platforms

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

N/A