Fit first media correctly after launching RV (#228)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added a new bool var in Session.cpp to check for a certain event
(before-progressive-loading) that happens before the
"after-progressive-loading" event.
The m_postFirstNonEmptyRender variable used to delay the window fitting
calls would always be called after launching RV, because an
"after-progressive-loading" event is launched even if there are no media
loaded (see RvApplication.cpp L:744).
The "before-progressive-loading" event is now used to prepare to check
for the "after-progressive-loading" event. Because the former should
only be called when media is wanting to get loaded.
There could have been a possibility to remove the
"after-progressive-loading" event call in RvApplication.cpp, but since
Session.cpp is not the only listener/receiver of events, it didn't feel
wise to do.
### Describe the reason for the change.
The "Fit Window to First media" checkbox wouldn't work when we loaded
the media after launching RV.
### Describe what you have tested and on which operating system.
MacOS
### Add a list of changes, and note any that might need special
attention during the review.
- Session.cpp/.h
### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>