Added Configurable FFMpeg build (#233)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
The FFMpeg external build now uses mostly CMake list vars for its
configure step.
### Describe the reason for the change.
We need to easily add new flags to the FFMpeg build, so that we can
modify the latter depending on our needs (in our case, commercial RV,
but the same mechanism could also be used for any super project
consuming Open RV as a git submodule).
### Describe what you have tested and on which operating system.
MacOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>