Native MACOSX_DEPLOYMENT_TARGET support (#242)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

N/A

### Summarize your change.

Unset macOS deployment target related variables so everything relies on
the MACOSX_DEPLOYMENT_TARGET environment variable.

### Describe the reason for the change.

Prior to this change, the macOS deployment targets were all over the
place because all third parties had it defined in different methods, now
we set the environment variable that the compiler understands so
everybody have it properly set.

### Describe what you have tested and on which operating system.

Tested on MacOS Ventura. Without MACOSX_DEPLOYMENT_TARGET, `otool -l
path/to/lib | grep minos` returns 13.0, and setting
MACOSX_DEPLOYMENT_TARGET=11.0 gives me 11.0. (excluding Qt which is
precompiled)

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>