Add possibility to change windows .exe icon (#232)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

### Summarize your change.
Added an IF(...) ELSE() statement when configuring the Windows rv.exe
icon.
### Describe the reason for the change.
You can now set an icon outside of the Open RV repo by targeting a .rc
file and assigning it to the RV_RC_ICON cmake var. This is necessary to
fix the Windows rv.exe icon showing itself as Open RV when it should be
the commercial one.
### Describe what you have tested and on which operating system.
Windows, we're only using the .rc file on Windows
### Add a list of changes, and note any that might need special
attention during the review.
- src/bin/apps/rv CMakeLists.txt
### If possible, provide screenshots.

Signed-off-by: ThomBern <thomas.bernier@autodesk.com>